correçoes code climate objCalculator.js

+++ src/objCalculator.js
@@ -23,7 +23,7 @@ const calculator = {};
 calculator.add = (param, param2) => {
   let sum = 0;
   if (typeof (param) === 'number' && typeof (param2) === 'number') {
-    sum = Math.floor(parseInt(param + param2));
+    sum = Math.floor(parseFloat(param + param2));
   }
   return sum; 
 }
@@ -31,7 +31,7 @@ calculator.add = (param, param2) => {
 calculator.div = (param, param2) => {
   let div;
   if (typeof (param) === 'number' && typeof (param2) === 'number') {
-    div = Math.floor(parseInt(param / param2));
+    div = Math.floor(parseFloat(param / param2));
   }
   return div;
 }
@@ -39,7 +39,7 @@ calculator.div = (param, param2) => {
 calculator.mult = (param, param2) => {
   let mult;
   if (typeof (param) === 'number' && typeof (param2) === 'number') {
-    mult = Math.floor(parseInt(param * param2));
+    mult = Math.floor(parseFloat(param * param2));
   }
   return mult;
 }
@@ -47,9 +47,8 @@ calculator.mult = (param, param2) => {
 calculator.sub = (param, param2) => {
   let sub;
   if (typeof (param) === 'number' && typeof (param2) === 'number') {
-    sub = Math.floor(parseInt(param - param2));
+    sub = Math.floor(parseFloat(param - param2));
   }
   return sub;
 }
-
 module.exports = calculator;
